Moved messages and tests of group module to sep file.
